requirements.txt
pywhatkit==5.4
python-dotenv==1.0.0 